pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "learning_llms_from_first_principles"
version = "0.1.0"
description = "Learning LLMs from First Principles"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.1",
    "tiktoken>=0.5.1",
    "jupyterlab>=4.0",
    "matplotlib>=3.7.1",
    "tensorflow>=2.15.0",
    "tqdm>=4.66.1",
    "numpy>=1.25,<2.0",
    "pandas>=2.2.1",
    "psutil>=5.9.5",
    "hydra-core>=1.3.2",
]

[project.scripts]
pretrain = "learning_llms_from_first_principles.trainer.train:main"
classify = "learning_llms_from_first_principles.trainer.classify:main"
instruct = "learning_llms_from_first_principles.trainer.instruct_finetuning:main"

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
lint = [
  "black --check --diff src tests",
  "isort --check --diff src tests",
  "ruff check src tests",
  "mypy src tests"
]
format = [
  "isort src tests",
  "black src tests",
  "ruff check --fix src tests"
]
all = [
  "lint",
  "test"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
disallow_any_generics = false

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
